Moved interface in a new package.
