+++ kura/org.eclipse.kura.api/META-INF/MANIFEST.MF
@@ -41,6 +41,7 @@ Export-Package: org.eclipse.kura;version="1.7.0",
  org.eclipse.kura.data.listener;version="1.0.1",
  org.eclipse.kura.data.transport.listener;version="1.0.1",
  org.eclipse.kura.db;version="2.0.0",
+ org.eclipse.kura.db.keyvalue;version="1.0.0",
  org.eclipse.kura.deployment.hook;version="1.0.0",
  org.eclipse.kura.driver;version="1.0.0",
  org.eclipse.kura.driver.descriptor;version="1.0.0",
