+++ kura/org.eclipse.kura.api/src/main/java/org/eclipse/kura/db/keyvalue/KeyValueDbService.java
@@ -1,4 +1,4 @@
-package org.eclipse.kura.db;
+package org.eclipse.kura.db.keyvalue;
 
 import org.eclipse.kura.KuraConnectException;
